Fixes for forthcoming release of URI
This is to deal with changes in the https://github.com/perl6-community-modules/uri/pull/44
which is planned to be released this week.  The only real changes are
that the query is now an object and needs to be coerced to a Hash in one
or two places.

I've tested with both as-is and forthcoming versions of URI to ensure it
works with both (the coercions will be a no-op in the as-is version)